Fix detection of whether StateArgs are expected by target state's OnEnter on certain compilers
In particular, gcc may fail to detect whether OnEnter expects StateArgs or not because it's able to optimize class layouts by stuffing child class data members into a parent class if there's room due to alignment padding. Because of this, comparing the sizes of the target class's Args against State::Args doesn't work, as they may have the same size, despite the target class defining its own derived type. To fix this, we just check if the two are actually the same type or not at compile time.