Add support for state overrides with state args
State args work by binding to a target state's OnEnter (with args) at compile time; but state overrides are only resolved at runtime by definition, so they cannot work with state args as is. My solution is to add transition function overloads for when a state override is passed with args, and to bind to the OnEnter call on the "source state" of the override. For this to work, the client must make the OnEnter a virtual function that is implemented in the source and all target override states. See the updated state_overrides sample for an example of how to use this.